backend/requirements-dev.txt
Django==3.0.8
djangorestframework
drf-nested-routers
Pillow==7.2.0
django-cors-middleware
django-webpack-loader
haikunator
djangorestframework-jwt
django-redis-cache